refactor: Extract summary private methods to mixin Module
